Add files via upload
New version of KrazyGlue that includes the new gluing of the metric as well as the calculation of X, Y, and the r- and t- partial derivatives of X and Y for all 3 regions (Schwarzchild, Friedman, and the Gluing Region). The calculation the derivatives was done both by hand and in Mathematica. The derivatives calculated using Mathematica were then converted from Wolfram Language code into Fortran(not kidding, but Mathematica has a built-in tool for that and it is really easy to convert Fortran to Python, just get rid of line continuation symbols) and then into Python.