Add files via upload
Changed fGenerator so it generates values of fear = f - r + 4/3 instead of f. Again, f and fbar are defined parametrically for large negative values of r, though the parametrization is different this time. Differentiation was changed to be done using finite difference methods. Since fbar' ~ 1 and fbar'' ~ 0 for r << 0, this should result in far faster runtimes as well as fewer numerical errors for a minimal loss of precision.